Adds interval filtering to Resource#query()
The commit also updates the montly vyew logic so that it
handles the new format.

By providing the property 'interval' with a DateInterval value
to the query function the result of the query can be limited
to the given date interval. By using the property name
'interval__X' where X is the name of an object property on the
server the filtering can be done on varios fields that support
gte/lte-filtering (see tastypie). X is 'order_date' by default.