Add datagrid widget for montly view
The datagrid is designed to be standalone but has only been
tested for the montly view use case. It displays data in a
bootstrap styled table.

The datagrid allows for CRUD operations on the data in the table.
Through callbacks the operations can be made to connect with a backend.